Add files via upload
This project involved analyzing COVID-19 data, specifically the number of deaths and cases, which was obtained from the website https://ourworldindata.org/covid-deaths. The first step was cleaning the data to ensure accuracy and reliability. After cleaning the data, the next step was analyzing the data using SQL queries. SQL queries were used to filter, group, and aggregate the data in various ways, allowing for a more in-depth analysis of the COVID-19 pandemic. Finally, a dashboard was created in Power BI to visualize the data in a clear and concise manner. The use of SQL and Power BI allowed for efficient data processing and visualization, making it easier to identify trends and patterns in the data. This project aimed to provide insights into the COVID-19 pandemic and to make the data more accessible and understandable for the general public. Overall, this project demonstrates the value of using data analysis tools like SQL and Power BI to gain insights into important global issues like the COVID-19 pandemic.