COVID-19_Data_Analysis_SQL_Queries.sql
This project involved analyzing COVID-19 data, specifically the number of deaths and cases, which was obtained from the website https://ourworldindata.org/covid-deaths. The data was then cleaned and analyzed using SQL queries. Finally, a dashboard was created in PowerBI using these queries to visualize the data in a clear and concise manner. This project aimed to provide insights into the COVID-19 pandemic and to make the data more accessible and understandable for the general public. The use of SQL and PowerBI allowed for efficient data processing and visualization, making it easier to identify trends and patterns in the data. Overall, this project demonstrates the value of using data analysis tools to gain insights into important global issues like the COVID-19 pandemic.
This is a link to Power BI: https://github.com/NMHDataAnalyst/Power-BI/blob/main/COVID-19%20Number%20of%20Deaths%20and%20Cases.pbix